% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_awl_ttr.R
\name{add_awl_ttr}
\alias{add_awl_ttr}
\title{Compute Average Word Length and Type-Token Ratio}
\usage{
add_awl_ttr(vec, ttr = 400)
}
\arguments{
\item{vec}{The vector of words for which to compute AWL and TTR.}

\item{ttr}{The number of words to consider for TTR calculation. Default is 400.}
}
\value{
A tibble containing the AWL and TTR values.
}
\description{
Computes the Average Word Length (AWL) and Type-Token Ratio (TTR) of a vector of words.
}
\examples{

add_awl_ttr(c("Hello world", "This is R"))

}
