% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mda_tags.R
\name{add_mda_tags}
\alias{add_mda_tags}
\title{Add MDA Tags}
\usage{
add_mda_tags(x, hesitation = TRUE, progress = FALSE)
}
\arguments{
\item{x}{character string}

\item{hesitation}{logical. If \code{TRUE}, hesitation markers are extracted from the text}

\item{progress}{logical. If \code{TRUE}, a progress message is printed}
}
\value{
character string
}
\description{
Adds MDA tags to a given text. The text must already include ST tags and tokenized.
}
\examples{
text <- c("I_PPSS", "have_VB", "a_DET", "dog_NN")
add_mda_tags(text)

}
