% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spell_suggestions.R
\name{add_spell_suggestions}
\alias{add_spell_suggestions}
\title{Identify Irregular Spellings and Suggest Corrections to a Table of Text}
\usage{
add_spell_suggestions(
  tbl,
  x = "text",
  window = 5,
  dict = "en_US",
  change_to = NULL
)
}
\arguments{
\item{tbl}{A data frame that contains the text data.}

\item{x}{A character string indicating the name of the text column in \code{tbl} (default is "text").}

\item{window}{An integer specifying the context window size for \code{quick_conc} function (default is 5).}

\item{dict}{A character string indicating the dictionary to be used by \code{hunspell} (default is "en_US").}

\item{change_to}{A character string specifying the column name ("alt_01" or "typo") to use for the \code{change_to} column (default is NULL).}
}
\value{
A tibble with identified spelling mistakes and corresponding suggested corrections.
Each row represents an instance of a spelling mistake, and each cell in the 'alt' column
represents a comma-separated list of spelling suggestions.

If \code{change_to} is specified, an additional column 'change_to' is included with
either the original incorrect spelling  (in case of \code{change_to = "typo"}), or the function's
best guess at the correct spelling (when \code{change_to = "alt_01"}).

This can be useful when exporting to a spreadsheet app for editing.
}
\description{
This function identifies spelling mistakes in a given text column of a data frame, suggests corrections,
and adds them as new columns. It uses the \code{hunspell} package for spell checking and providing suggestions.
The output can be exported to a spreadsheet for easier editing.
}
\examples{
\dontrun{
data <- tibble(text = c("This is a smaple text.", "Another txt with a typo."))
add_spell_suggestions(data)
}

}
