% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_st_tags.R
\name{add_st_tags}
\alias{add_st_tags}
\title{Add ST tags}
\usage{
add_st_tags(x, mdl = udmodel, hesitation = FALSE, tokenized = FALSE)
}
\arguments{
\item{x}{A character vector containing the text to be tagged.}

\item{mdl}{A model object of class \code{udpipe_model}.}

\item{hesitation}{Logical. Should hesitation markers be included?}

\item{tokenized}{Logical. Set to TRUE in case text has already been tokenized.}
}
\value{
A character vector containing the tagged text.
}
\description{
Adds Stanford (ST) part of speech tags to a character vector of text.
}
\details{
This function adds Stanford Tags (ST) to a vector of text. The function works by tokenizing the text first, and then annotating the tokenized text using \code{udpipe::udpipe_annotate}. Requires a udpipe_model to be present, so run init_udpipe_model() first.
}
