% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tag_tbl.R
\name{add_tag_tbl}
\alias{add_tag_tbl}
\title{Add a Table of ST, MDA and Other UD Tags to Text}
\usage{
add_tag_tbl(text, ...)
}
\arguments{
\item{text}{A character vector of input text to be processed.}

\item{...}{Additional arguments to be passed to the \code{add_st_tags()} function.}
}
\value{
A tibble with the original text annotated with ST tags and MDA tags. The output columns include \code{doc_id}, \code{st}, and \code{mda}.
}
\description{
The \code{add_tag_tbl} function is a wrapper around the \code{add_st_tags} function that also adds MDA tags to the text.
It processes and annotates the input text using the full Universal Dependencies (UD) model with the udpipe package, then adds MDA tags to the resulting output.
}
\examples{
\dontrun{
# Load the required udpipe model
init_udpipe_model()


# Process text with the add_tag_tbl function
text <- "This is a sample sentence."
result <- add_tag_tbl(text)
}
}
