% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_mda_tags.R
\name{dtag_negation}
\alias{dtag_negation}
\alias{dtag_prepositions}
\alias{dtag_ind_pron}
\alias{dtag_quantifiers}
\alias{dtag_quant_pron}
\title{Basic MDA Tags}
\usage{
dtag_negation(x)

dtag_prepositions(x)

dtag_ind_pron(x)

dtag_quantifiers(x)

dtag_quant_pron(x)
}
\arguments{
\item{x}{character vector of _ST tagged text, tokenized.}
}
\value{
x character vector with <MDA> tags appended.
}
\description{
Append basic <MDA> tags to a tokenized vector of _ST tagged strings.
}
\details{
This group of dtag functions are best applied before other dtags as subsequent ones may depend on them.
The functions match various case insensitive regex patterns and append the relevant
<MDA> tag to the end of the respective token in the vector.
}
\section{Functions}{
\itemize{
\item \code{dtag_negation}: MDA Tags tag negation

\item \code{dtag_prepositions}: MDA Tags tag preopositions

\item \code{dtag_ind_pron}: MDA Tags tag indefenite pronouns

\item \code{dtag_quantifiers}: MDA Tags tag quantifiers

\item \code{dtag_quant_pron}: MDA Tags tag quantifier pronouns
}}

\examples{
\dontrun{
# Generate some text
text <- "Anyone using this text should find some examples among these words.
Despite not each pattern being referenced, several are among the words,
and enough for somebody to get the gist."
# Add Stanford tags to text
text <- add_st_tags(text)
# dtag_negation
dtag_negation(text)
# dtag_prepostions
dtag_prepositions(text)
# dtag_ind_pron
dtag_ind_pron(text)
# dtag_quantifiers
dtag_quantifiers(text)
# dtag_quant_pron
dtag_quant_pron(text)}
}
