% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_mda_tags.R
\name{dtag_negation}
\alias{dtag_negation}
\alias{dtag_prepositions}
\alias{dtag_ind_pron}
\alias{dtag_quantifiers}
\alias{dtag_quant_pron}
\title{basicTagging

A collection of functions to add basic <MDA> tags to a tokenized vector of ST- tagged strings

This collection of tags do SOMETHING
#' basicTagging}
\usage{
dtag_negation(x)

dtag_prepositions(x)

dtag_ind_pron(x)

dtag_quantifiers(x)

dtag_quant_pron(x)
}
\arguments{
\item{x}{character vector of ST tagged text tokenized}
}
\description{
basicTagging

A collection of functions to add basic <MDA> tags to a tokenized vector of ST- tagged strings

This collection of tags do SOMETHING
#' basicTagging
}
\section{Functions}{
\itemize{
\item \code{dtag_negation}: tag negation

\item \code{dtag_prepositions}: tag preopositions

\item \code{dtag_ind_pron}: tag indefenite pronouns

\item \code{dtag_quantifiers}: tag quantifiers

\item \code{dtag_quant_pron}: tag quantifier pronouns
}}

