% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tags.R
\name{compare_tags}
\alias{compare_tags}
\title{Compare Tags}
\usage{
compare_tags(vec1, vec2, regex = NULL, regex1 = NULL, regex2 = NULL, ...)
}
\arguments{
\item{vec1}{A character vector of tagged text in tokenised form used for reference.}

\item{vec2}{A character vector of tagged text in tokenised form to compare against vec1.}

\item{regex}{A regular expression of the tag to be checked in both vectors.}

\item{regex1}{The regular expression of the tag to search in vector 1.}

\item{regex2}{The regular expression of the tag to search in vector 2.}

\item{...}{Additional arguments to be passed on to the \code{quick_conc} output.}
}
\value{
A tibble showing the \code{quick_conc} concordance of the tags in
vector 1 displaid line by line in comparison to vector 2.
}
\description{
This function compares the tags between two vectors of strings.
}
\details{
This function is used to compare the tagging results om two texts.
The texts should be reduced to two vectors in tokenised form. The function
produces concordance lines of the tags as chosen by regex or index position.
Both inputs should be exactly the same length.
}
