% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_mda_tags.R
\name{dtag_pred_adj}
\alias{dtag_pred_adj}
\alias{dtag_emphatics}
\alias{dtag_phrasal_coord}
\alias{dtag_pro_do}
\alias{dtag_wh_questions}
\alias{dtag_sentence_rels}
\alias{dtag_perfect_asp}
\alias{dtag_passives_by}
\alias{dtag_be_main}
\alias{dtag_wh_clauses}
\alias{dtag_pp_rel_clauses}
\alias{dtag_str_prepositions}
\alias{dtag_split_infinitives}
\alias{dtag_split_auxiliaries}
\alias{dtag_syn_negation}
\alias{dtag_time_adverbials}
\alias{dtag_place_adverbials}
\alias{dtag_that_vc}
\alias{dtag_that_ac}
\alias{dtag_pres_part}
\alias{dtag_past_part}
\alias{dtag_past_whiz}
\alias{dtag_pres_whiz}
\alias{dtag_that_subj}
\alias{dtag_that_obj}
\alias{dtag_wh_subj}
\alias{dtag_wh_obj}
\alias{dtag_hedges}
\alias{dtag_disc_part}
\alias{dtag_dem_pronouns}
\alias{dtag_demonstratives}
\alias{dtag_that_del}
\alias{dtag_indep_cc}
\title{Complex MDA Tags}
\usage{
dtag_pred_adj(x)

dtag_emphatics(x)

dtag_phrasal_coord(x)

dtag_pro_do(x)

dtag_wh_questions(x)

dtag_sentence_rels(x)

dtag_perfect_asp(x)

dtag_passives_by(x)

dtag_be_main(x)

dtag_wh_clauses(x)

dtag_pp_rel_clauses(x)

dtag_str_prepositions(x)

dtag_split_infinitives(x)

dtag_split_auxiliaries(x)

dtag_syn_negation(x)

dtag_time_adverbials(x)

dtag_place_adverbials(x)

dtag_that_vc(x)

dtag_that_ac(x)

dtag_pres_part(x)

dtag_past_part(x)

dtag_past_whiz(x)

dtag_pres_whiz(x)

dtag_that_subj(x)

dtag_that_obj(x)

dtag_wh_subj(x)

dtag_wh_obj(x)

dtag_hedges(x)

dtag_disc_part(x)

dtag_dem_pronouns(x)

dtag_demonstratives(x)

dtag_that_del(x)

dtag_indep_cc(x)
}
\arguments{
\item{x}{character vector of _ST tagged text, tokenized.}
}
\value{
x character vector with <MDA> tags appended.
}
\description{
Append complex <MDA> tags to a tokenized vector of _ST tagged strings.
}
\details{
This group of dtag functions are best applied after the Basic  MDA tags as they may rely on them.
The functions match various case insensitive regex patterns and append the relevant
<MDA> tag to the end of the respective token in the vector.
}
\section{Functions}{
\itemize{
\item \code{dtag_pred_adj}: MDA Tags tag predicative adjectives

\item \code{dtag_emphatics}: MDA Tags tag emphatics

\item \code{dtag_phrasal_coord}: MDA Tags tag phrasal "and" coordination

\item \code{dtag_pro_do}: MDA Tags tag pro-verb do

\item \code{dtag_wh_questions}: MDA Tags tag WH questions

\item \code{dtag_sentence_rels}: MDA Tags tag sentence relatives

\item \code{dtag_perfect_asp}: MDA Tags tag perfect aspects

\item \code{dtag_passives_by}: MDA Tags tag passives with by

\item \code{dtag_be_main}: MDA Tags tag be as main verb

\item \code{dtag_wh_clauses}: MDA Tags tag WH clauses

\item \code{dtag_pp_rel_clauses}: MDA Tags pied piper relative clauses

\item \code{dtag_str_prepositions}: MDA Tags tag stranded Prepositions

\item \code{dtag_split_infinitives}: MDA Tags tag Split Infinitive

\item \code{dtag_split_auxiliaries}: MDA Tags tag Split Auxiliaries

\item \code{dtag_syn_negation}: MDA Tags tag Synthetic Negation

\item \code{dtag_time_adverbials}: MDA Tags tag Time Adverbials

\item \code{dtag_place_adverbials}: MDA Tags tag Place Adverbials

\item \code{dtag_that_vc}: MDA Tags tag 'that' verb complements

\item \code{dtag_that_ac}: MDA Tags tag 'that' adjective complements

\item \code{dtag_pres_part}: MDA Tags tag Present Participial Clauses

\item \code{dtag_past_part}: MDA Tags tag Past Participial Clauses

\item \code{dtag_past_whiz}: MDA Tags tag Past Participial WHIZ Deletion Relatives

\item \code{dtag_pres_whiz}: MDA Tags tag Present Participial WHIZ Deletion Relatives

\item \code{dtag_that_subj}: MDA Tags tag "that" Relative Clauses on Subject Position

\item \code{dtag_that_obj}: MDA Tags tag "that" Relative Clauses on Object Position

\item \code{dtag_wh_subj}: MDA Tags tag WH Relative Clauses on Subject Position

\item \code{dtag_wh_obj}: MDA Tags tag WH Relative Clauses on Object Position

\item \code{dtag_hedges}: MDA Tags tag Hedges

\item \code{dtag_disc_part}: MDA Tags tag Discourse Particles

\item \code{dtag_dem_pronouns}: MDA Tags tag Demonstrative Pronouns

\item \code{dtag_demonstratives}: MDA Tags tag Demonstratives

\item \code{dtag_that_del}: MDA Tags tag Subordinator-That Deletion

\item \code{dtag_indep_cc}: MDA Tags tag Independent Clause Coordination
}}

\examples{
\dontrun{
# Generate some text
text <- "This example is short and sweet. This means that not all the complex
tags will have been included, which is why this is really only a guide and it should
be used with that in mind. Otherwise, I think you will be disappointed."
# Load udpipe model into the global environment for _ST tagging
init_udpipe_model()
# Add Stanford tags to text
text <- add_st_tags(text)
# dtag_passives
dtag_passives(text)
# dtag_that_del
dtag_that_del(text)
# dtag_demonstratives
dtag_demonstratives(text)
# dtag_dem_pronouns
dtag_dem_pronouns(text)
# dtag_sentence_rels
dtag_sentence_rels(text)
# dtag_emphatics
dtag_emphatics(text)
#dtag_be_main
dtag_be_main(text)
#dtag_perfect_asp
dtag_perfect_asp(text)
# #dtag_pred_adj
dtag_pred_adj(text)}
}
