% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_mda_tags.R
\name{dtag_pred_adj}
\alias{dtag_pred_adj}
\alias{dtag_emphatics}
\alias{dtag_phrasal_coord}
\alias{dtag_pro_do}
\alias{dtag_wh_questions}
\alias{dtag_sentence_rels}
\alias{dtag_perfect_asp}
\alias{dtag_passives}
\alias{dtag_passives_by}
\alias{dtag_be_main}
\alias{dtag_wh_clauses}
\alias{dtag_pp_rel_clauses}
\alias{dtag_str_prepositions}
\alias{dtag_split_infinitives}
\alias{dtag_split_auxiliaries}
\alias{dtag_syn_negation}
\alias{dtag_time_adverbials}
\alias{dtag_place_adverbials}
\alias{dtag_that_vc}
\alias{dtag_that_ac}
\alias{dtag_pres_part}
\alias{dtag_past_part}
\alias{dtag_past_whiz}
\alias{dtag_pres_whiz}
\alias{dtag_that_subj}
\alias{dtag_that_obj}
\alias{dtag_wh_subj}
\alias{dtag_wh_obj}
\alias{dtag_hedges}
\alias{dtag_disc_part}
\alias{dtag_dem_pronouns}
\alias{dtag_demonstratives}
\alias{dtag_that_del}
\alias{dtag_indep_cc}
\title{complexTagging}
\usage{
dtag_pred_adj(x)

dtag_emphatics(x)

dtag_phrasal_coord(x)

dtag_pro_do(x)

dtag_wh_questions(x)

dtag_sentence_rels(x)

dtag_perfect_asp(x)

dtag_passives(x, by = TRUE)

dtag_passives_by(x)

dtag_be_main(x)

dtag_wh_clauses(x)

dtag_pp_rel_clauses(x)

dtag_str_prepositions(x)

dtag_split_infinitives(x)

dtag_split_auxiliaries(x)

dtag_syn_negation(x)

dtag_time_adverbials(x)

dtag_place_adverbials(x)

dtag_that_vc(x)

dtag_that_ac(x)

dtag_pres_part(x)

dtag_past_part(x)

dtag_past_whiz(x)

dtag_pres_whiz(x)

dtag_that_subj(x)

dtag_that_obj(x)

dtag_wh_subj(x)

dtag_wh_obj(x)

dtag_hedges(x)

dtag_disc_part(x)

dtag_dem_pronouns(x)

dtag_demonstratives(x)

dtag_that_del(x)

dtag_indep_cc(x)
}
\arguments{
\item{x}{Character vector of _ST tagged text, tokenized.}
}
\value{
Character vector of tokenized text with <MDA> tags appended
}
\description{
complexTagging
}
\section{Functions}{
\itemize{
\item \code{dtag_pred_adj}: tag predicative adjectives

\item \code{dtag_emphatics}: tag emphatics

\item \code{dtag_phrasal_coord}: tag phrasal "and" coordination

\item \code{dtag_pro_do}: tag pro-verb do

\item \code{dtag_wh_questions}: tag WH questions

\item \code{dtag_sentence_rels}: tag sentence relatives

\item \code{dtag_perfect_asp}: tag perfect aspects

\item \code{dtag_passives}: tag passives

\item \code{dtag_passives_by}: tag passives with by

\item \code{dtag_be_main}: tag be as main verb

\item \code{dtag_wh_clauses}: tag WH clauses

\item \code{dtag_pp_rel_clauses}: pied piper relative clauses

\item \code{dtag_str_prepositions}: tag stranded Prepositions

\item \code{dtag_split_infinitives}: tag Split Infinitive

\item \code{dtag_split_auxiliaries}: tag Split Auxiliaries

\item \code{dtag_syn_negation}: tag Synthetic Negation

\item \code{dtag_time_adverbials}: tag Time Adverbials

\item \code{dtag_place_adverbials}: tag Place Adverbials

\item \code{dtag_that_vc}: tag 'that' verb complements

\item \code{dtag_that_ac}: tag 'that' adjective complements

\item \code{dtag_pres_part}: tag Present Participial Clauses

\item \code{dtag_past_part}: tag Past Participial Clauses

\item \code{dtag_past_whiz}: tag Past Participial WHIZ Deletion Relatives

\item \code{dtag_pres_whiz}: tag Present Participial WHIZ Deletion Relatives

\item \code{dtag_that_subj}: tag "that" Relative Clauses on Subject Position

\item \code{dtag_that_obj}: tag "that" Relative Clauses on Object Position

\item \code{dtag_wh_subj}: tag WH Relative Clauses on Subject Position

\item \code{dtag_wh_obj}: tag WH Relative Clauses on Object Position

\item \code{dtag_hedges}: tag Hedges

\item \code{dtag_disc_part}: tag Discourse Particles

\item \code{dtag_dem_pronouns}: tag Demonstrative Pronouns

\item \code{dtag_demonstratives}: tag Demonstratives

\item \code{dtag_that_del}: tag Subordinator-That Deletion

\item \code{dtag_indep_cc}: tag Independent Clause Coordination
}}

