% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_flatten_text.R
\name{d_flatten_text}
\alias{d_flatten_text}
\title{Flatten Text}
\usage{
d_flatten_text(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector where the elements are a single string.
}
\description{
This function takes a vector of strings and flattens them into a single text.
}
\details{
A wrapper for stringr::str_flatten() with the additional function of
removing floating spaces between a set of punctuation marks (.,;:!?) and tokens.
}
\examples{
text <- c("This" , "is" , "," , "or" , "at" , "least" , "should" , "be" , "," ,
"a" , "suitable" , "example" , "." , "Is" , "it", "?")
d_flatten_text(text)
}
