% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_sub.R
\name{d_sub}
\alias{d_sub}
\title{Substitute Patterns in a Vector}
\usage{
d_sub(x, pattern, replacement, perl = TRUE, ...)
}
\arguments{
\item{x}{A vector.}

\item{pattern}{The pattern to look for.}

\item{replacement}{The replacement string.}

\item{perl}{Logical value. If \code{TRUE}, pattern is a Perl-compatible regular expression. If \code{FALSE}, pattern is a POSIX basic regular expression.}

\item{...}{Further arguments passed to \code{\link[base]{sub}}.}
}
\value{
A vector with patterns replaced.
}
\description{
Substitute patterns in a vector \code{x} with a given \code{replacement},  using base::sub for speed, by stringr format for convenience
}
\examples{
d_sub(x = c("abc", "def", "ghi"), pattern = "[a-z]{3}", replacement = "0")

}
