% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Tukey_hsd.R
\name{d_Tukey_hsd}
\alias{d_Tukey_hsd}
\title{Run Tukey HSD on Dimension scores}
\usage{
d_Tukey_hsd(df, dimension = "dimension", corpus = "corpus")
}
\arguments{
\item{df}{A data frame of dimension scores in long format, with a column for
identifying the input corpus.}

\item{dimension}{The dimension for the Tukey HSD test.}

\item{corpus}{The corpus for the Tukey HSD test.}
}
\value{
A data frame with the Tukey HSD test results.
}
\description{
This function performs Tukey HSD to a data frame of dimension scores where more
than one corpus have been anaylsed. If using the result of \code{dtag_tbl},
the data frame needs to be changed to a long format, e.g. by using the
\code{pivot_longer(tbl , contains("Dimension"), names_to = "dimension",
names_transform = list(dimension = as.factor))}, with an extra column to identify
the source corpus.
}
