% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_mda_tags.R
\name{dtag_amplifier}
\alias{dtag_amplifier}
\alias{dtag_downtoner}
\alias{dtag_nominalisation}
\alias{dtag_gerund}
\alias{dtag_all_nouns}
\alias{dtag_all_adjectives}
\alias{dtag_all_adverbs}
\alias{dtag_present_tenses}
\alias{dtag_first_person_pronoun}
\alias{dtag_second_person_pronoun}
\alias{dtag_third_person_pronoun}
\alias{dtag_pronoun_it}
\alias{dtag_causative}
\alias{dtag_concessive}
\alias{dtag_conditional}
\alias{dtag_possibility_modal}
\alias{dtag_necessity_modal}
\alias{dtag_predictive_modal}
\alias{dtag_public_verb}
\alias{dtag_private_verb}
\alias{dtag_suasive_verb}
\alias{dtag_seem_appear}
\alias{dtag_contractions}
\title{Dimension Tags by Regex Patterns}
\usage{
dtag_amplifier(x)

dtag_downtoner(x)

dtag_nominalisation(x)

dtag_gerund(x)

dtag_all_nouns(x)

dtag_all_adjectives(x)

dtag_all_adverbs(x)

dtag_present_tenses(x)

dtag_first_person_pronoun(x)

dtag_second_person_pronoun(x)

dtag_third_person_pronoun(x)

dtag_pronoun_it(x)

dtag_causative(x)

dtag_concessive(x)

dtag_conditional(x)

dtag_possibility_modal(x)

dtag_necessity_modal(x)

dtag_predictive_modal(x)

dtag_public_verb(x)

dtag_private_verb(x)

dtag_suasive_verb(x)

dtag_seem_appear(x)

dtag_contractions(x)
}
\arguments{
\item{x}{character vector of _ST tagged text, tokenized.}
}
\value{
x character vector with <MDA> tags appended.
}
\description{
Append <MDA> tags after pattern matching with regex.
}
\details{
This group of dtag functions are best applied after other dtags have been processed.
The functions match various case insensitive regex patterns and append the relevant
<MDA> tag to the end of the respective token in the vector.
}
\section{Functions}{
\itemize{
\item \code{dtag_amplifier}: Tags by Regex Patterns tag amplifiers

\item \code{dtag_downtoner}: Tags by Regex Patterns tag downtoners

\item \code{dtag_nominalisation}: Tags by Regex Patterns tag nominalisations

\item \code{dtag_gerund}: Tags by Regex Patterns tag gerunds

\item \code{dtag_all_nouns}: Tags by Regex Patterns tag for all nouns

\item \code{dtag_all_adjectives}: Tags by Regex Patterns tag for all attributive adjectives

\item \code{dtag_all_adverbs}: Tags by Regex Patterns tag for all adverbs

\item \code{dtag_present_tenses}: Tags by Regex Patterns tag for all present tenses

\item \code{dtag_first_person_pronoun}: Tags by Regex Patterns tag first person pronouns

\item \code{dtag_second_person_pronoun}: Tags by Regex Patterns tag second person pronouns

\item \code{dtag_third_person_pronoun}: Tags by Regex Patterns tag third person pronouns

\item \code{dtag_pronoun_it}: Tags by Regex Patterns tag pronoun it

\item \code{dtag_causative}: Tags by Regex Patterns tag causative because

\item \code{dtag_concessive}: Tags by Regex Patterns tag concessive subordinators

\item \code{dtag_conditional}: Tags by Regex Patterns tag conditional subordinators

\item \code{dtag_possibility_modal}: Tags by Regex Patterns tag possibility modals

\item \code{dtag_necessity_modal}: Tags by Regex Patterns tag necessity modals

\item \code{dtag_predictive_modal}: Tags by Regex Patterns tag predictive modals

\item \code{dtag_public_verb}: Tags by Regex Patterns tag public verbs

\item \code{dtag_private_verb}: Tags by Regex Patterns tag private verbs

\item \code{dtag_suasive_verb}: Tags by Regex Patterns tag suasive verbs

\item \code{dtag_seem_appear}: Tags by Regex Patterns tag seem/appear

\item \code{dtag_contractions}: Tags by Regex Patterns tag contractions
}}

\examples{
 \dontrun{
# Generate some text
text <- "This text is merely a quick exemplification of some dtag options.
Because it is short, you shouldn't expect very much."
# Add Stanford tags to text
text <- add_st_tags(text)
# dtag_amplifier
dtag_amplifier(text)
# dtag_downtoner
dtag_downtoner(text)
# dtag_nominalistation
dtag_nominalisation(text)
# dtag_all_adjectives
dtag_all_adjectives(text)
# dtag_second_person_pronoun
dtag_second_person_pronoun(text)
# dtag_pronoun_it
dtag_pronoun_it(text)
# dtag_necessity modal
dtag_necessity modal(text)
# dtag_contractions
dtag_contractions(text)}
}
