% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_adverbial_subords.R
\name{dtag_adverbial_subords}
\alias{dtag_adverbial_subords}
\title{Tag Adverbial Subordinating Conjunctions}
\usage{
dtag_adverbial_subords(x)
}
\arguments{
\item{x}{A character vector of strings that have been tokenized and tagged with _ST tags.}
}
\value{
A character vector with adverbial subordinating conjunctions <OSUB> tags added.
}
\description{
This function tags adverbial subordinating conjunctions in text that has been previously _ST tagged and tokenized.
}
\examples{
dtag_adverbial_subords(c("Since_RB",  "he_PRP" , "left_VBD"))


}
