% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_all_adjectives.R
\name{dtag_all_adjectives}
\alias{dtag_all_adjectives}
\title{Add adjective <JJ> tag to all attributive adjectives}
\usage{
dtag_all_adjectives(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with all attributive adjectives tagged by <JJ>.
}
\description{
Adds the generic adjective tag  <JJ> to all attributive adjectives
based on regex match.
}
