% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_all_adverbs.R
\name{dtag_all_adverbs}
\alias{dtag_all_adverbs}
\title{Add adverb <RB> tag to all adverbs}
\usage{
dtag_all_adverbs(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some adverbs tagged by <RB>.
}
\description{
Adds the generic adverb tag  <RB> to all adverbs not included in
other categories.
}
\details{
The functions adds the <RB> tag to all adverbs except the following:
\itemize{
\item Amplifiers <AMP>
\item Discourse particles <DPAR>
\item Downtoners <DWNT>
\item Emphatis <EMPH>
\item Subordinating conjunctions <OSUB>
\item Place adverbials <PLACE>
\item Quantifiers <QUAN>
\item Time adverbials <TIME>
\item Not <XX0>
}
}
