% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_all_nouns.R
\name{dtag_all_nouns}
\alias{dtag_all_nouns}
\title{Add noun <NN> tag to all nouns}
\usage{
dtag_all_nouns(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with  all nouns tagged by <NN>.
}
\description{
Adds the generic noun tag  <NN> to all nouns based on regex match.
}
