% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_all_pp.R
\name{dtag_all_pp}
\alias{dtag_all_pp}
\title{Add past participle <VBN> tag to all past participle verb forms}
\usage{
dtag_all_pp(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with all past participle verb matches tagged by <VBN>.
}
\description{
Adds the generic past participle  tag  <VBN> to allpast participle verb forms
based on regex match.
}
