% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_be_main.R
\name{dtag_be_main}
\alias{dtag_be_main}
\title{Add BE as main verb <BEMA> tag}
\usage{
dtag_be_main(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some BE as main verb <BEMA> tags.
}
\description{
Adds the BE as main verb tag <BEMA> based on a regex match.
}
