% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_causative.R
\name{dtag_causative}
\alias{dtag_causative}
\title{Add causative <CAUS> tag}
\usage{
dtag_causative(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some causative  <CAUS> tags.
}
\description{
Adds the causative tag <CAUS> based on a regex match.
}
\details{
The function matches both "because" and the informal "cos" as causatives.
}
