% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_concessive.R
\name{dtag_concessive}
\alias{dtag_concessive}
\title{Add concessive subordinator <CONC> tag}
\usage{
dtag_concessive(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some concessive subordinator  <CONC> tags.
}
\description{
Adds the concessive subordinator tag <CONC> based on a regex match.
}
\details{
The function matches "although" and both "though" and the shortened
form "tho" as concessive subordinators.
}
