% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_conditional.R
\name{dtag_conditional}
\alias{dtag_conditional}
\title{Add conditional subordinator <COND> tag}
\usage{
dtag_conditional(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some conditional subordinator  <COND> tags.
}
\description{
Adds the conditional subordinator tag <COND> based on a regex match.
}
