% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_contractions.R
\name{dtag_contractions}
\alias{dtag_contractions}
\title{Add contractions <CONT> tag}
\usage{
dtag_contractions(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some contraction <CONT> tags.
}
\description{
Adds the contraction tag <CONT> based on a regex match.
}
