% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_dem_pronouns.R
\name{dtag_dem_pronouns}
\alias{dtag_dem_pronouns}
\title{Add Demonstrative Pronouns <DEMP> tag}
\usage{
dtag_dem_pronouns(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some demonstrative pronouns <DEMP> tags.
}
\description{
Adds demonstrative pronouns <DEMP> based on regex matches.
}
