% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_demonstratives.R
\name{dtag_demonstratives}
\alias{dtag_demonstratives}
\title{Add Demonstratives  <DEMO> tag}
\usage{
dtag_demonstratives(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some demonstratives <DEMO> tags.
}
\description{
Adds demonstratives <DEMO> based on a regex match.
}
