% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_disc_part.R
\name{dtag_disc_part}
\alias{dtag_disc_part}
\title{Add discourse particles <DPART> tag}
\usage{
dtag_disc_part(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some discourse particles <DPART> tags.
}
\description{
Adds the discourse particles tag <DPART> based on a regex match.
}
