% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_downtoner.R
\name{dtag_downtoner}
\alias{dtag_downtoner}
\title{Tag downtoners}
\usage{
dtag_downtoner(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some downtoner <DWNT> tags.
}
\description{
Adds the downtoner tag  <DWNT> based on a regex match.
}
