% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_emphatics.R
\name{dtag_emphatics}
\alias{dtag_emphatics}
\title{Add emphatic <EMPH> tag}
\usage{
dtag_emphatics(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some emphatic <EMPH> tags.
}
\description{
Adds the emphatic tag <EMPH> based on a regex match.
}
\details{
In case of multiword units such as "a lot", the second token has
its _ST tag replaced with _NULL
}
