% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_ex_there.R
\name{dtag_ex_there}
\alias{dtag_ex_there}
\title{Tag existential there}
\usage{
dtag_ex_there(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some existential there <EX> tags.
}
\description{
Adds the existential there tag  <EX> based on a regex match.
}
