% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_first_person_pronoun.R
\name{dtag_first_person_pronoun}
\alias{dtag_first_person_pronoun}
\title{Add first person pronoun <FPP1> tag}
\usage{
dtag_first_person_pronoun(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some first person pronoun  <FPP1> tags.
}
\description{
Adds the first person pronoun tag <FPP1> based on a regex match.
}
