% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_gerund.R
\name{dtag_gerund}
\alias{dtag_gerund}
\title{Tag gerunds}
\usage{
dtag_gerund(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some gerund <GER> tags.
}
\description{
Adds the gerund tag  <GER> based on a regex match.
}
