% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_hedges.R
\name{dtag_hedges}
\alias{dtag_hedges}
\title{Add hedges <HDG> tag}
\usage{
dtag_hedges(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some emphatic <EMPH> tags.
}
\description{
Adds the hedges <HDG> based on a regex match.
}
\details{
In case of multiword units such as "more or less", the second and third
token has its _ST tag replaced with _NULL.
}
