% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_hesitation.R
\name{dtag_hesitation}
\alias{dtag_hesitation}
\title{Add hesitation markers}
\usage{
dtag_hesitation(x, regex = "\\\\berm?\\\\b|\\\\bum\\\\b")
}
\arguments{
\item{x}{A character vector.}

\item{regex}{A regular expression (default \\berm?\\b|\\bum\\b).}
}
\value{
A character vector with hesitation tags.
}
\description{
Add hesitation marker <HSTN> to a vector of tokenized strings
}
\examples{
dtag_hesitation(c("I'm", "not", "sure",".", "Um" ,"," ,"no"))

}
