% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_indep_cc.R
\name{dtag_indep_cc}
\alias{dtag_indep_cc}
\title{Add Independent Clause Coordination <ANDC> tag}
\usage{
dtag_indep_cc(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some independent clause coordination <ANDC> tags.
}
\description{
Adds independent clause coordination <ANDC> based on regex matches.
}
