% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_necessity_modal.R
\name{dtag_necessity_modal}
\alias{dtag_necessity_modal}
\title{Add necessity modal <NEMD> tag}
\usage{
dtag_necessity_modal(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some necessity modal   <NEMD> tags.
}
\description{
Adds the necessity modal tag <NEMD> based on a regex match.
}
