% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_negation.R
\name{dtag_negation}
\alias{dtag_negation}
\title{Tag analytic negation}
\usage{
dtag_negation(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some negation <XX0> tags.
}
\description{
Adds the analytic negation tag  <XX0> based on a regex match.
}
