% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_nominalisation.R
\name{dtag_nominalisation}
\alias{dtag_nominalisation}
\title{Tag nominalisation}
\usage{
dtag_nominalisation(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some nominalisation <NOMZ> tags.
}
\description{
Adds the nominalisation tag  <NOMZ> based on a regex match.
}
