% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_passives_by.R
\name{dtag_passives_by}
\alias{dtag_passives_by}
\title{Add passive tags <BYPA>}
\usage{
dtag_passives_by(x)
}
\arguments{
\item{x}{A character vector of strings that have been tokenized and tagged with _ST tags.}
}
\value{
A character vector with passive <BYPA> tags added.
}
\description{
This function tags by passives in text that has been previously _ST
tagged, tokenized and tagged by the  dtag_passives() function.
}
\details{
The function replaces the <PASS> tag with <BYPA> if a "by" token
is found within 1 to 3 tokens of the tagged passive.

Usually there is no need to run this function unless the \code{dtag_passives(by = FALSE)}
has been run first.
}
