% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_past_part.R
\name{dtag_past_part}
\alias{dtag_past_part}
\title{Add past participial <PASTP> tag}
\usage{
dtag_past_part(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some past participial <PASTP> tags.
}
\description{
Adds past participial <PASTP> based on regex matches.
}
