% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_past_tenses.R, R/dtag_perfect_tenses.R
\name{dtag_past_tenses}
\alias{dtag_past_tenses}
\title{Add past tense <VBD> tag to all verbs in past tense}
\usage{
dtag_past_tenses(x)

dtag_past_tenses(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with all past tense verb matches  tagged by <VBD>.

A character vector with all perfect form matches  tagged by <VBN>.
}
\description{
Adds the generic past tense tag  <VBD> to all past tenses
based on regex match.

Adds the genericperfect form <VBN> to all perfect verb forms
based on regex match.
}
