% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_past_whiz.R
\name{dtag_past_whiz}
\alias{dtag_past_whiz}
\title{Add past participial WHIZ deletion relative <WZPAST> tag}
\usage{
dtag_past_whiz(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some past participial WHIZ deletion relative <WZPAST> tags.
}
\description{
Adds past participial WHIZ deletion relative <WZPAST> based on regex matches.
}
