% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_perfect_asp.R
\name{dtag_perfect_asp}
\alias{dtag_perfect_asp}
\title{Add perfect aspect <PEAS> tag}
\usage{
dtag_perfect_asp(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some prefect aspect <PEAS> tags.
}
\description{
Adds the perfect aspect tag <PEAS> based on a regex match.
}
