% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_phrasal_coord.R
\name{dtag_phrasal_coord}
\alias{dtag_phrasal_coord}
\title{Add phrasal coordination <PHC> tag}
\usage{
dtag_phrasal_coord(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some phrasal coordination <PHC> tags.
}
\description{
Adds the phrasal coordination tag <PHC> based on a regex match.
}
\details{
The <PHC> tag is added in case of "and" with certain conditions
such as:
\itemize{
\item between two adverb tags
\item between two adjective tags
\item between two noun tags
}
}
