% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_place_adverbials.R
\name{dtag_place_adverbials}
\alias{dtag_place_adverbials}
\title{Add place adverbial <PLACE> tag}
\usage{
dtag_place_adverbials(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some place adverbial <PLACE> tags.
}
\description{
Adds place adverbial <PLACE> based on a regex match.
}
