% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_possessives.R
\name{dtag_possessives}
\alias{dtag_possessives}
\title{Correct Stanford (ST) possessive tags}
\usage{
dtag_possessives(x)
}
\arguments{
\item{x}{a character vector of strings, tokenized and with _ST tags.}
}
\value{
a character vector of strings, tokenized and with "..$" replaced by "..S"
}
\description{
Correct Stanford _ST possessive tags by replacing $ with S to make future regex matching easier.
}
\examples{
x <- c("Their_PRP$", "house_NN", "was_VBD", "big_JJ", "and_CC", "luxurious_JJ", "._.")
dtag_possessives(x)
}
