% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_possibility_modal.R
\name{dtag_possibility_modal}
\alias{dtag_possibility_modal}
\title{Add possibility modal <POMD> tag}
\usage{
dtag_possibility_modal(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some possibility modal   <POMD> tags.
}
\description{
Adds the possibility modal tag <POMD> based on a regex match.
}
