% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_pp_rel_clauses.R
\name{dtag_pp_rel_clauses}
\alias{dtag_pp_rel_clauses}
\title{Add pied piper relative clause <PIRE> tag}
\usage{
dtag_pp_rel_clauses(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some pied piper relative clause <PIRE> tags.
}
\description{
Adds the pied piper relative clausetag <PIRE> based on a regex match.
}
