% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_predictive_modal.R
\name{dtag_predictive_modal}
\alias{dtag_predictive_modal}
\title{Add predictive modal <PRMD> tag}
\usage{
dtag_predictive_modal(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some predictive modal   <PRMD> tags.
}
\description{
Adds the predictive modal tag <PRMD> based on a regex match.
}
