% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drag_prepositions.R
\name{dtag_prepositions}
\alias{dtag_prepositions}
\title{Add preposition <PIN> tag}
\usage{
dtag_prepositions(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some preposition <PIN> tags.
}
\description{
Adds the preposition tag  <PIN> based on a regex match.
}
