% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_pres_part.R
\name{dtag_pres_part}
\alias{dtag_pres_part}
\title{Add present participial <PRESP> tag}
\usage{
dtag_pres_part(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some present participial <PRESP> tags.
}
\description{
Adds present participial <PRESP> based on regex matches.
}
