% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_pres_whiz.R
\name{dtag_pres_whiz}
\alias{dtag_pres_whiz}
\title{Add present participial WHIZ deletion relative <WZPRES> tag}
\usage{
dtag_pres_whiz(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some present participial WHIZ deletion relative <WZPRES> tags.
}
\description{
Adds present participial WHIZ deletion relative <WZPRES> based on regex matches.
}
