% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_present_tenses.R
\name{dtag_present_tenses}
\alias{dtag_present_tenses}
\title{Add present tense <VPRT> tag to all verbs in present tense}
\usage{
dtag_present_tenses(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with all present tense verb matches  tagged by <VPRT>.
}
\description{
Adds the generic present tense tag  <VPRT> to all present tenses
based on regex match.
}
