% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_private_verb.R
\name{dtag_private_verb}
\alias{dtag_private_verb}
\title{Add private verb <PRIV> tag}
\usage{
dtag_private_verb(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some private verb <PRIV> tags.
}
\description{
Adds the private verb tag <PRIV> based on a regex match.
}
\details{
For a list of the private verbs matched, see the shorthand character vector
named private \code{sh["private"]}
}
