% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_pro_do.R
\name{dtag_pro_do}
\alias{dtag_pro_do}
\title{Add pro verb do <PROD> tag}
\usage{
dtag_pro_do(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some pro verb do <PROD> tags.
}
\description{
Adds the pro verb do tag <PROD> based on a regex match.
}
