% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_pronoun_it.R
\name{dtag_pronoun_it}
\alias{dtag_pronoun_it}
\title{Add it pronoun <PIT> tag}
\usage{
dtag_pronoun_it(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some it pronoun  <PIT> tags.
}
\description{
Adds the it pronoun tag <PIT> based on a regex match.
}
