% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_public_verb.R
\name{dtag_public_verb}
\alias{dtag_public_verb}
\title{Add public verb <PUBV> tag}
\usage{
dtag_public_verb(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some public verb <PUBV> tags.
}
\description{
Adds the public verb tag <PUBV> based on a regex match.
}
\details{
For a list of the public verbs matched, see the shorthand character vector
named public \code{sh["public"]}
}
