% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_quant_pron.R
\name{dtag_quant_pron}
\alias{dtag_quant_pron}
\title{Tag quantifier pronouns}
\usage{
dtag_quant_pron(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some quantifier pronoun <QUPR> tags.
}
\description{
Adds the quantifier pronoun tag  <QUPR> based on a regex match.
}
