% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_quantifiers.R
\name{dtag_quantifiers}
\alias{dtag_quantifiers}
\title{Tag quantifiers}
\usage{
dtag_quantifiers(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some quantifier <QUAN> tags.
}
\description{
Adds the quantifier tag  <QUAN> based on a regex match.
}
