% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_second_person_pronoun.R
\name{dtag_second_person_pronoun}
\alias{dtag_second_person_pronoun}
\title{Add second person pronoun <SPP2> tag}
\usage{
dtag_second_person_pronoun(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some second person pronoun  <SPP2> tags.
}
\description{
Adds the second person pronoun <SPP2> based on a regex match.
}
