% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_seem_appear.R
\name{dtag_seem_appear}
\alias{dtag_seem_appear}
\title{Add seem / appear verb <SMP> tag}
\usage{
dtag_seem_appear(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some seem / appear  verb <SMP> tags.
}
\description{
Adds the seem / appear  verb tag <SMP> based on a regex match.
}
