% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_sentence_rels.R
\name{dtag_sentence_rels}
\alias{dtag_sentence_rels}
\title{Add sentence relatives <SERE> tag}
\usage{
dtag_sentence_rels(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some sentence relatives <SERE> tags.
}
\description{
Adds the sentence relatives tag <SERE> based on a regex match.
}
