% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_split_auxiliaries.R
\name{dtag_split_auxiliaries}
\alias{dtag_split_auxiliaries}
\title{Add split auxiliary <SPAU> tag}
\usage{
dtag_split_auxiliaries(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some split auxiliary  <SPAU> tags.
}
\description{
Adds split auxiliary <SPAU> based on a regex match.
}
