% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_str_prepositions.R
\name{dtag_str_prepositions}
\alias{dtag_str_prepositions}
\title{Add stranded preposition <STPR> tag}
\usage{
dtag_str_prepositions(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some stranded preposition  <STPR> tags.
}
\description{
Adds stranded preposition  <STPR> based on a regex match.
}
