% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_suasive_verb.R
\name{dtag_suasive_verb}
\alias{dtag_suasive_verb}
\title{Add suasive verb <SUAV> tag}
\usage{
dtag_suasive_verb(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some suasive verb <SUAV> tags.
}
\description{
Adds the suasive verb tag <SUAV> based on a regex match.
}
\details{
For a list of the suasive verbs matched, see the shorthand character vector
named suasive \code{sh["suasive"]}
}
