% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_syn_negation.R
\name{dtag_syn_negation}
\alias{dtag_syn_negation}
\title{Add synthetic negation <SYNE> tag}
\usage{
dtag_syn_negation(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some synthetic negation <SYNE> tags.
}
\description{
Adds synthetic negation <SYNE> based on a regex match.
}
