% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_tbl.R
\name{dtag_tbl}
\alias{dtag_tbl}
\title{dtag_tbl}
\usage{
dtag_tbl(tbl, input = 1, text = 2, ttr = 400)
}
\arguments{
\item{tbl}{A data frame with at least one column as input id  and another for the _ST tagged text.}

\item{input}{A column for the input id (defaults to 1st position, but can be named as e.g. "colname1").}

\item{text}{A column for the text (defaults to 2nd position, but can be named as e.g. "colname12").
The text should be tagged with _ST tags, and in the flattened, not tokenized form.}

\item{ttr}{Maximum number of tokens to consider for TTR, defaults to 400.}
}
\value{
A tibble containing:
\itemize{
\item wordcount - number of non-punctuation tokens found in text
\item dimension - Dimension1 ~ Dimension6 from Biber 1988 for each feature
\item feature - the <MDA> tag or AWL or TTR
\item detail - brief description of the feature
\item count - number of times the feature is counted in text
\item value - in case of <MDA> tag, normailsed frequency per 100 tokens
\item z-score - value scaled to the biber_mean and biber_sd
\item d-score - same as z-score, but with the sign of negative dimension features reversed
\item biber_mean and biber_sd for each feature, based on Biber 1988
\item closest matching text type for each input, based on Biber 1989
}
}
\description{
Compute the Biber-style dimension scores of a data frame.
}
