% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_that_ac.R
\name{dtag_that_ac}
\alias{dtag_that_ac}
\title{Add THAT as adjectival complement <THAC> tag}
\usage{
dtag_that_ac(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some THAT as adjectival complement <THAC> tags.
}
\description{
Adds THAT as adjectival complement <THAC> based on a regex match.
}
