% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_that_del.R
\name{dtag_that_del}
\alias{dtag_that_del}
\title{Add "That" Deletion <THATD> tag}
\usage{
dtag_that_del(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some "that" deletion <THATD> tags.
}
\description{
Adds "that" deletion <THATD> based on regex matches.
}
