% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_that_obj.R
\name{dtag_that_obj}
\alias{dtag_that_obj}
\title{Add "that" Relative Clauses on Object Position  <TOBJ> tag}
\usage{
dtag_that_obj(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some "that" relative clauses on object Position <TOBJ> tags.
}
\description{
Adds "that" relative clauses on object position <TOBJ> based on regex matches.
}
