% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_that_subj.R
\name{dtag_that_subj}
\alias{dtag_that_subj}
\title{Add "that" Relative Clauses on Subject Position  <TSUB> tag}
\usage{
dtag_that_subj(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some "that" relative clauses on subject position <TSUB> tags.
}
\description{
Adds "that" relative clauses on subject position <TSUB> based on regex matches.
}
