% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_that_vc.R
\name{dtag_that_vc}
\alias{dtag_that_vc}
\title{Add THAT as verbal complement <THVC> tag}
\usage{
dtag_that_vc(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some THAT as verbal complement <THVC> tags.
}
\description{
Adds THAT as verbal complement <THVC> based on regex matches.
}
