% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_third_person_pronoun.R
\name{dtag_third_person_pronoun}
\alias{dtag_third_person_pronoun}
\title{Add third person pronoun <TPP3> tag}
\usage{
dtag_third_person_pronoun(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some third person pronoun  <TPP3> tags.
}
\description{
Adds the third person pronoun tag <TPP3> based on a regex match.
}
