% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_time_adverbials.R
\name{dtag_time_adverbials}
\alias{dtag_time_adverbials}
\title{Add time adverbial <TIME> tag}
\usage{
dtag_time_adverbials(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some time adverbial <TIME> tags.
}
\description{
Adds time adverbial <TIME> based on a regex match.
}
