% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_to_inf.R
\name{dtag_to_inf}
\alias{dtag_to_inf}
\title{Tag infinitive to}
\usage{
dtag_to_inf(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with infinitive to tag  <TO> added.
}
\description{
Adds the infinitive to tag  <TO> to all to tokens based on a regex match,
}
\details{
Adds the infinitive to tag  <TO> to all tokens of the form "to_".
The function should be followed by the dtag_to_prep() function to distinguish
to when used as a preposition.
}
