% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_to_prep.R
\name{dtag_to_prep}
\alias{dtag_to_prep}
\title{Correction of TO as preposition}
\usage{
dtag_to_prep(x)
}
\arguments{
\item{x}{A character of tokenized strings with ST tags}
}
\value{
A character vector with some preposition tags.
}
\description{
Adds the preposition tag  <PIN> to preposition form of to_TO
}
