% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_wh_clauses.R
\name{dtag_wh_clauses}
\alias{dtag_wh_clauses}
\title{Add Wh- clause <WHCL> tag}
\usage{
dtag_wh_clauses(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some Wh- clause <WHCL> tags.
}
\description{
Adds the Wh- clause tag <WHCL> based on a regex match.
}
