% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_wh_obj.R
\name{dtag_wh_obj}
\alias{dtag_wh_obj}
\title{Add WH Relative Clauses on Object Position <WHOBJ> tag}
\usage{
dtag_wh_obj(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some WH relative clauses on object position <WHOBJ> tags.
}
\description{
Adds WH relative clauses on object position <WHOBJ> based on regex matches.
}
