% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_wh_questions.R
\name{dtag_wh_questions}
\alias{dtag_wh_questions}
\title{Add WH- question <WHQU> tag}
\usage{
dtag_wh_questions(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some WH- question <WHQU> tags.
}
\description{
Adds the WH- question tag <WHQU> based on a regex match.
}
