% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_wh_subj.R
\name{dtag_wh_subj}
\alias{dtag_wh_subj}
\title{Add WH Relative Clauses on Subject Position <WHSUB> tag}
\usage{
dtag_wh_subj(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with some WH relative clauses on subject Position <WHSUB> tags.
}
\description{
Adds WH relative clauses on subject position <WHSUB> based on regex matches.
}
