% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtagger-package.R
\docType{package}
\name{dtagger-package}
\alias{dtagger}
\alias{dtagger-package}
\title{dtagger: Suite of Functions to Tag Text for Multidimensional Analysis}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timmarchand/dtagger}
  \item \url{https://timmarchand.github.io/dtagger/}
  \item Report bugs at \url{https://github.com/timmarchand/dtagger/issues}
}

}
\author{
\strong{Maintainer}: Tim Marchand \email{timmarchand@gmail.com}

}
\keyword{internal}
