% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Find the mode of a given vector}
\usage{
find_mode(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The mode of the vector
}
\description{
This function takes a vector and finds the mode (most frequent value)
}
\examples{
find_mode(c(1,2,2,2,3,4,4))
# 2
}
