% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_non_ascii.R
\name{find_non_ascii}
\alias{find_non_ascii}
\title{Find Non-ASCII Characters in Text}
\usage{
find_non_ascii(tbl, id = NULL, text = "text", sort_by_chr = FALSE, ...)
}
\arguments{
\item{tbl}{A data frame that contains the text data.}

\item{id}{A character string indicating the name of the identifier column in \code{tbl} (default is NULL).}

\item{text}{A character string indicating the name of the text column in \code{tbl} (default is "text").}

\item{sort_by_chr}{A logical indicating whether to sort the results by the non-ASCII characters (default is FALSE).}

\item{...}{Arguments to pass on to \code{quick_conc()}. For example, you can extend the
resulting search window witb the argument \code{n = 10},}
}
\value{
A tibble with identified non-ASCII characters. Each row represents an instance of a non-ASCII
character. If \code{id} is not NULL, the tibble also includes the identifier for each instance.
If \code{sort_by_chr} is TRUE, the tibble is sorted by the non-ASCII characters.
}
\description{
This function finds non-ASCII characters in a given text column of a data frame. It uses
the \code{quick_conc} function for finding non-ASCII characters. The results can optionally be sorted
by the non-ASCII characters.
}
\examples{
\dontrun{
data <- tibble(id = 1:2,
text = c("This is a text with a non-ASCII character: é.", "Another text without."))
find_non_ascii(data, id = "id")
}

}
