% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_mda_tags.R
\name{dtag_amplifier}
\alias{dtag_amplifier}
\alias{dtag_downtoner}
\alias{dtag_nominalisation}
\alias{dtag_gerund}
\alias{dtag_all_nouns}
\alias{dtag_all_adjectives}
\alias{dtag_all_adverbs}
\alias{dtag_present_tenses}
\alias{dtag_first_person_pronoun}
\alias{dtag_second_person_pronoun}
\alias{dtag_third_person_pronoun}
\alias{dtag_pronoun_it}
\alias{dtag_causative}
\alias{dtag_concessive}
\alias{dtag_conditional}
\alias{dtag_possibility_modal}
\alias{dtag_necessity_modal}
\alias{dtag_predictive_modal}
\alias{dtag_public_verb}
\alias{dtag_private_verb}
\alias{dtag_suasive_verb}
\alias{dtag_seem_appear}
\alias{dtag_contractions}
\title{groupedTagging}
\usage{
dtag_amplifier(x)

dtag_downtoner(x)

dtag_nominalisation(x)

dtag_gerund(x)

dtag_all_nouns(x)

dtag_all_adjectives(x)

dtag_all_adverbs(x)

dtag_present_tenses(x)

dtag_first_person_pronoun(x)

dtag_second_person_pronoun(x)

dtag_third_person_pronoun(x)

dtag_pronoun_it(x)

dtag_causative(x)

dtag_concessive(x)

dtag_conditional(x)

dtag_possibility_modal(x)

dtag_necessity_modal(x)

dtag_predictive_modal(x)

dtag_public_verb(x)

dtag_private_verb(x)

dtag_suasive_verb(x)

dtag_seem_appear(x)

dtag_contractions(x)
}
\arguments{
\item{x}{character vector of ST tagged text, tokenized.}
}
\description{
groupedTagging
}
\section{Functions}{
\itemize{
\item \code{dtag_amplifier}: tag amplifiers

\item \code{dtag_downtoner}: tags downtoners

\item \code{dtag_nominalisation}: tags nominalisations

\item \code{dtag_gerund}: tags gerunds

\item \code{dtag_all_nouns}: tags for all nouns

\item \code{dtag_all_adjectives}: tags for all attributive adjectives

\item \code{dtag_all_adverbs}: tags for all adverbs

\item \code{dtag_present_tenses}: tags for all present tenses

\item \code{dtag_first_person_pronoun}: tags first person pronouns

\item \code{dtag_second_person_pronoun}: tags second person pronouns

\item \code{dtag_third_person_pronoun}: tags third person pronouns

\item \code{dtag_pronoun_it}: tags pronoun it

\item \code{dtag_causative}: tags causative because

\item \code{dtag_concessive}: tags concessive subordinators

\item \code{dtag_conditional}: tags conditional subordinators

\item \code{dtag_possibility_modal}: tags possibility modals

\item \code{dtag_necessity_modal}: tags necessity modals

\item \code{dtag_predictive_modal}: tags predictive modals

\item \code{dtag_public_verb}: tags public verbs

\item \code{dtag_private_verb}: tags private verbs

\item \code{dtag_suasive_verb}: tags suasive verbs

\item \code{dtag_seem_appear}: tags seem/appear

\item \code{dtag_contractions}: tags contractions
}}

