% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_udpipe_model.R
\name{init_udpipe_model}
\alias{init_udpipe_model}
\title{Initialize Model}
\usage{
init_udpipe_model(lang = "english")
}
\arguments{
\item{lang}{The language to use for the model. Default is 'english'.}
}
\value{
None
}
\description{
Initialize Model
}
\examples{
init_udpipe_model()

}
