% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_midpipe.R
\name{log_midpipe}
\alias{log_midpipe}
\title{Log Midpipe}
\usage{
log_midpipe(x, expr)
}
\arguments{
\item{x}{A data object.}

\item{expr}{An expression to be evaluated on the data object.}
}
\value{
The data object after the expression is evaluated.
}
\description{
A function to interrupt a pipe to print an expression before continuing with the pipe.
}
