% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_tags.R
\name{missing_tags}
\alias{missing_tags}
\title{Find Missing Tags}
\usage{
missing_tags(vec1, vec2, regex = NULL, regex1 = NULL, regex2 = NULL, ...)
}
\arguments{
\item{vec1}{A character vector of tagged text in tokenised form used for reference.}

\item{vec2}{A character vector of tagged text in tokenised form to compare against vec1.}

\item{regex}{A regular expression of the tag to be checked in both vectors.}

\item{regex1}{The regular expression of the tag to search in vector 1.}

\item{regex2}{The regular expression of the tag to search in vector 2.}

\item{...}{Additional arguments to be passed on to the \code{quick_conc} output.}
}
\value{
A tibble showing the \code{quick_conc} concordance of the missing tags in
vector 2 in comparison to vector 1.
}
\description{
This function finds the tags missing from one vector and present in another.
}
\details{
This function is used to compare the tagging results on the same text.
The texts should be reduced to two vectors in tokenised form. The function
produces concordance lines of where tags in vector one are missing from the same
index position in vector two. Both inputs should be exactly the same length.
}
