% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_replace_all.R
\name{multi_replace_all}
\alias{multi_replace_all}
\title{Replace multiple text tokens in strings}
\usage{
multi_replace_all(text, tbl, add_breaks = TRUE)
}
\arguments{
\item{text}{A string vector to replace tokens in.}

\item{tbl}{A two-column tibble with the \code{token} to replace and its
corresponding \code{new} value.}

\item{add_breaks}{A logical indicating whether to add \verb{\\\\b} (word boundary) to token.}
}
\value{
A character vector of strings with replaced tokens.
}
\description{
Replace multiple text tokens with new text in provided string.
}
\details{
The function
}
\examples{
# Replace multiple tokens
text <- c("Hello world!", "Goodbye world!")
tbl <- data.frame(token = c("Hello", "world"), new = c("Hi", "universe"))
multi_replace_all(text, tbl)
#> [1] "Hi universe!" "Goodbye universe!"

}
