% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_conc.R
\name{quick_conc}
\alias{quick_conc}
\title{Quick concordancing of pattern matches or index locations}
\usage{
quick_conc(x, index, n = 5, tokenize = FALSE, separated = FALSE)
}
\arguments{
\item{x}{a character vector of tokenized strings, or a single string}

\item{index}{a character vector of regex pattern to match, or a numeric vector to use as index of matches}

\item{n}{an integer, to specify the number of context tokens either side of the matched node}

\item{tokenize}{a logical, to tokenize the text first or not}

\item{separated}{a logical, to separate the context tokens or not}
}
\value{
A tibble containing:
\itemize{
\item case - a case number for the match found.
\item left - objects immediately adjacent (up to n) to the left of the matched node.
In case of \code{separated = TRUE}, the left  are separated into left(n):left1
\item match - the matched search item, as defined by the \code{index} argument.
\item right - tokens immediately adjacent (up to n) to the right of the matched node.
In case of \code{separated = TRUE}, the right tokens are separated into right1:right(n).
\item index - the index row position of matched result from the input data frame.
}
}
\description{
Lightweight concordancing function to return key words in context (KWIC) in a tidy format.
}
\examples{
x <- c("The", "cat", "sat", "on", "the", "mat")
index <- c("cat", "sat")
quick_conc(x, index, n = 2)
x <- "The dog barked loudly, alerting the neighbors of potential danger.
A nearby park seemed like the perfect spot for the dog and
it quickly made its way there."
quick_conc(x, index = "dog", n = 3, tokenize = TRUE, separated = TRUE)
quick_conc(x, index = c(4,8,12), tokenize = TRUE)

}
