% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_tags.R
\name{remove_duplicated_tags}
\alias{remove_duplicated_tags}
\title{Removes duplicated tags}
\usage{
remove_duplicated_tags(x)
}
\arguments{
\item{x}{A vector of tags}
}
\value{
A vector of tags with duplicates removed
}
\description{
Removes duplicated tags
}
