% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh.R
\docType{data}
\name{sh}
\alias{sh}
\title{Shorthand Reference Data}
\format{
A named vector of regex patterns
}
\usage{
sh
}
\description{
Shorthand for frequently used regex in tagging functions. The regex patterns can be matched by calling sh["name_of_pattern"].
}
\details{
Data and shortcuts used in some of the tagging functions
}
\keyword{datasets}
