% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_tag.R
\name{tokenize_tag}
\alias{tokenize_tag}
\title{Tokenize tag}
\usage{
tokenize_tag(vec, regex = "\\\\s(?!<\\\\w+>)")
}
\arguments{
\item{vec}{A character vector.}

\item{regex}{The regular expression used for splitting the vector. Use the default
setting for <MDA> tags.}
}
\value{
A list of character vectors split by tag.
}
\description{
Split a character vector based on a tag. The function tokenizes MDA
tagged texts by splitting on each space not followed by an <MDA> tag.
It will also work on _ST tags by default.
}
