% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mda_tags.R
\name{add_mda_tags}
\alias{add_mda_tags}
\title{Add <MDA> Tags}
\usage{
add_mda_tags(x, mda_hesitation = TRUE, progress = FALSE, ...)
}
\arguments{
\item{x}{character string}

\item{mda_hesitation}{logical. If \code{TRUE}, hesitation markers are extracted from the text.
Experimental feature - should hesitation markers be excluded before tagging?
Regex for the hesitation markers is the same as the default for \code{dtag_hesitation},
but can be set using the regex argument. See \code{\link{dtag_hesitation}} for details.}

\item{progress}{logical. If \code{TRUE}, a progress message is printed}

\item{...}{Additional arguments to pass on.}
}
\value{
character string with <MDA> tags added
}
\description{
Adds <MDA> tags to a given text. The text must already include _ST tags and tokenized.
}
\examples{
\dontrun{
# Simple example:
text <- c("I_PPSS", "have_VB", "a_DET", "dog_NN")
add_mda_tags(text)
# More detailed example:
# Generate some text
text <- "This example is short and sweet. This means that not all the tags will
have been included, which is why this is really only a guide and it should
be used with that in mind. Otherwise, I think it may lead to disappointment."
# Load udpipe model into the global environment for _ST tagging
init_udpipe_model()
# Add Stanford tags to text
text <- add_st_tags(text)
# Add <MDA> tags
add_mda_tags(text)}
}
