% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mda_tags.R
\name{add_mda_tags}
\alias{add_mda_tags}
\title{Add <MDA> Tags}
\usage{
add_mda_tags(x, mda_hesitation = TRUE, progress = FALSE, ...)
}
\arguments{
\item{x}{character string}

\item{mda_hesitation}{logical. If \code{TRUE}, hesitation markers are extracted from the text.
Experimental feature - should hesitation markers be excluded before tagging?
Regex for the hesitation markers is the same as the default for \code{dtag_hesitation}, but can be set using the regex argument. See \code{\link{dtag_hesitation}} for details.}

\item{progress}{logical. If \code{TRUE}, a progress message is printed}

\item{...}{Additional arguments to pass on.}
}
\value{
character string with <MDA> tags added
}
\description{
Adds <MDA> tags to a given text. The text must already include _ST tags and tokenized.
}
\examples{
text <- c("I_PPSS", "have_VB", "a_DET", "dog_NN")
add_mda_tags(text)

}
