% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_st_tags.R
\name{add_st_tags}
\alias{add_st_tags}
\title{Add Stanford _ST tags}
\usage{
add_st_tags(x, mdl = udmodel, st_hesitation = FALSE, tokenized = FALSE, ...)
}
\arguments{
\item{x}{A character vector containing the text to be tagged.}

\item{mdl}{A model object of class \code{udpipe_model}.}

\item{st_hesitation}{Logical. Experimental feature - should hesitation markers be excluded before tagging? Set to TRUE to exclude.
Regex for the hesitation markers are the same as the default for \code{dtag_hesitation}, but can be set using the regex argument. See \code{\link{dtag_hesitation}} for details.}

\item{tokenized}{Logical. Set to TRUE in case text has already been tokenized.}

\item{...}{Additional arguments to pass on.}
}
\value{
A character vector containing the tagged text.
}
\description{
Adds Stanford (_ST) part of speech tags to a character vector of text.
}
\details{
This function adds Stanford Tags _ST to a vector of text. The function works by tokenizing the text first,
and then annotating the tokenized text using \code{udpipe::udpipe_annotate}.
Requires a udpipe_model to be present, so if an error occurs, try running \code{init_udpipe_model()} first.
}
