% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_folder.R
\name{dtag_folder}
\alias{dtag_folder}
\title{dtag_folder}
\usage{
dtag_folder(path, n = NULL)
}
\arguments{
\item{path}{A character string denoting the folder containing the text files.}

\item{n}{An optional argument denoting the maximum number of text files to be analyzed.}
}
\value{
A list of dataframes containing the tagged texts, individual and corpus-level scores for each dimension of the text, and word counts.
}
\description{
This function takes in a path and an optional argument for the number of texts to be analyzed. It then reads the text files in the folder, adds stylistic and multidimensional analysis tags, and returns a list of dataframes containing the tagged texts, individual and corpus-level scores for each dimension of the text, and word counts.
}
\examples{
\dontrun{
dtag_folder("path_to_folder")}
}
