% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_hesitation.R
\name{dtag_hesitation}
\alias{dtag_hesitation}
\title{Add hesitation markers}
\usage{
dtag_hesitation(x, regex = "\\\\berm?\\\\b|\\\\berm?_|\\\\bum\\\\b|\\\\bum_")
}
\arguments{
\item{x}{A character vector with <HSTN> added, where necessary.}

\item{regex}{A regular expression (default "\\berm?\\b|\\berm?\emph{|\\bum\\b|\\bum}").
The regex expression is case insensitive by default.}
}
\value{
A character vector with hesitation tags appended.
}
\description{
Add hesitation marker <HSTN> to a vector of tokenized strings.
}
\examples{
dtag_hesitation(c("I'm", "not", "sure",".", "Um" ,"," ,"no"))

}
