% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtag_all_pp.R, R/dtag_past_tenses.R
\name{dtag_past_tenses}
\alias{dtag_past_tenses}
\title{Add past participle <VBN> tag to all past participle verb forms}
\usage{
dtag_past_tenses(x)

dtag_past_tenses(x)
}
\arguments{
\item{x}{A character of tokenized strings with _ST tags.}
}
\value{
A character vector with all past participle verb matches tagged by <VBN>.

A character vector with all past tense verb matches  tagged by <VBD>.
}
\description{
Adds the generic past participle  tag  <VBN> to allpast participle verb forms
based on regex match.

Adds the generic past tense tag  <VBD> to all past tenses
based on regex match.
}
