% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_flatten_text.R
\name{d_flatten}
\alias{d_flatten}
\alias{d_flatten_text}
\title{Flatten Text}
\usage{
d_flatten(x)

d_flatten_text(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector where the elements are a single string.
}
\description{
This function takes a vector of strings and flattens them into a single string.
}
\details{
A wrapper for stringr::str_flatten(x, " ").

In case of d_flatten_text(), there is the additional function of removing
floating spaces between a set of punctuation marks (.,;:!?") and tokens.
}
\section{Functions}{
\itemize{
\item \code{d_flatten}: d_flatten

\item \code{d_flatten_text}: d_flatten_text
}}

\examples{
x <- c("This" , "is" , "," , "or" , "at" , "least" , "should" , "be" , "," ,
"a" , "suitable" , "example" , "." , "Is" , "it", "?")
d_flatten(x)
d_flatten_text(x)

}
