% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_udpipe_model.R
\name{init_udpipe_model}
\alias{init_udpipe_model}
\title{Initialize UDPipe Model}
\usage{
init_udpipe_model(lang = "english-gum")
}
\arguments{
\item{lang}{The language to use for the model. Default is 'english-gum'.}
}
\value{
None
}
\description{
Initialize UDPipe Model
}
\examples{
\dontrun{init_udpipe_model()}

}
